Fix bug in burn, sleep and stun effect Change statusName to EffectName Change statusValue to EffectValue Add RoundEvent and Effect for positions only

+++ Components/Statistics.cs
@@ -18,7 +18,7 @@
         public int maxMana;
         public float movingSpeed = 0.75f;
         public double shield;
-        public float specialAtkPercentage = 1;
+        public double specialAtkPercentage = 1;
         public int specialDef;
 
         public double hp
