+++ Effects/Stun.cs
@@ -9,11 +9,11 @@ namespace TFT_Engine.Effects
 {
     class Stun : Effect
     {
-        public Stun(int Duration, Character effector, Character effected) : base(Duration, effector, effected)
+        public Stun(double Duration, Character effector, Character effected) : base(Duration, effector, effected)
         {
         }
 
-        public Stun(int Duration, Set effector, Character effected) : base(Duration, effector, effected)
+        public Stun(double Duration, Set effector, Character effected) : base(Duration, effector, effected)
         {
         }
     }
